+++ base/strings/util.jl
@@ -190,14 +190,10 @@ julia> rstrip(a)
 ```
 """
 function rstrip(s::AbstractString, chars::Chars=_default_delims)
-    r = RevString(s)
-    i = start(r)
-    while !done(r,i)
-        c, j = next(r,i)
-        if !(c in chars)
-            return SubString(s, 1, endof(s)-i+1)
-        end
-        i = j
+    i = endof(s)
+    while i > 0
+        s[i] in chars || return SubString(s, 1, i)
+        i = prevind(s, i)
     end
     SubString(s, 1, 0)
 end